requirements.txt
aiohttp==3.8.1
aiohttp-socks==0.7.1
aiorpcX==0.18.7
aiosignal==1.2.0
anyio==3.3.4
anytree==2.8.0
apparmor==3.0.3
appdirs==1.4.4
asgiref==3.4.1
astroid==2.8.3
async-timeout==4.0.1
attrs==21.2.0
Babel==2.9.1
Beaker==1.11.0
bitstring==3.1.9
btrfsutil==5.15.1
CacheControl==0.12.6
certifi==2021.10.8
cffi==1.14.6
chardet==4.0.0
charset-normalizer==2.0.4
click==8.0.3
colorama==0.4.4
contextlib2==0.6.0.post1
cryptography==35.0.0
distlib==0.3.3
distro==1.6.0
Django==3.2.9
djangorestframework==3.12.4
djangorestframework-simplejwt==5.0.0
dnspython==2.1.0
docopt==0.6.2
ecdsa==0.17.0
Electrum==4.1.5
filelock==3.0.12
Flask==2.0.2
Flask-Babel==2.0.0
frozenlist==1.2.0
gevent==21.8.0
greenlet==1.1.2
h11==0.12.0
h2==4.0.0
hpack==4.0.0
html5lib==1.1
httpcore==0.13.7
httpx==0.19.0
httpx-socks==0.5.2
hyperframe==6.0.1
idna==3.3
importlib-metadata==4.8.1
isort==5.9.3
itsdangerous==2.0.1
Jinja2==3.0.3
jsonrpclib-pelix==0.4.3.1
langdetect==1.0.9
lazy-object-proxy==1.6.0
LibAppArmor==3.0.3
libfdt==1.6.1
libtorrent===1.2.14-build-libtorrent-rasterbar-src-libtorrent-rasterbar-1.2.14-bindings-python
libvirt-python==7.8.0
lxml==4.6.4
Mako==1.1.6
Markdown==3.3.6
MarkupSafe==2.0.1
mccabe==0.6.1
meson==0.60.2
more-itertools==8.10.0
msgpack==1.0.2
multidict==5.2.0
ordered-set==4.0.2
packaging==20.9
pbkdf2==1.3
pep517==0.11.0
Pillow==8.4.0
platformdirs==2.4.0
ply==3.11
progress==1.6
protobuf==3.17.3
psutil==5.8.0
pyaes==1.6.1
pycairo==1.20.1
pycparser==2.20
pycryptodomex==3.10.1
Pygments==2.10.0
PyGObject==3.42.0
PyJWT==2.3.0
pylint==2.11.1
pyOpenSSL==21.0.0
pyparsing==2.4.7
PyQt5==5.15.6
PyQt5-sip==12.9.0
PySocks==1.7.1
python-dateutil==2.8.2
python-socks==2.0.0
python-xlib==0.31
pytz==2021.3
PyYAML==5.4.1
qrcode==7.3
ranger-fm==1.9.3
Reflector==2021.7.8.19.52.52
requests==2.26.0
resolvelib==0.5.5
retrying==1.3.3
rfc3986==1.5.0
screenkey==1.4
searx==1.0.0
six==1.16.0
sniffio==1.2.0
speaklater==1.3
sqlparse==0.4.2
team==1.0
toml==0.10.2
tomli==1.2.2
trash-cli==0.21.10.24
typing-extensions==3.10.0.2
ueberzug==18.1.9
urllib3==1.26.7
virtualenv==20.8.0
webencodings==0.5.1
websocket-client==0.59.0
Werkzeug==2.0.2
wrapt==1.13.2
yapf==0.31.0
yarl==1.7.2
youtube-dl==2021.6.6
zipp==3.6.0
zope.event==4.5.0
zope.interface==5.4.0
